Reworked value sizeof logic to (1) properly treat user types with size, (2) properly treat self-called _sizeof, avoiding fake UserType wrapping

+++ shared/src/main/scala/io/kaitai/struct/precompile/CalculateSeqSizes.scala
@@ -19,7 +19,7 @@ object CalculateSeqSizes {
             sizeElement
           case RepeatExpr(expr) =>
             expr.evaluateIntConst match {
-              case Some(count) => FixedSized(elementSize * count)
+              case Some(count) => FixedSized(elementSize * count.toInt)
               case None => DynamicSized
             }
           case _: RepeatUntil | RepeatEos =>
@@ -108,7 +108,7 @@ object CalculateSeqSizes {
       case FloatMultiType(width, _) => FixedSized(width.width)
       case _: BytesEosType => DynamicSized
       case blt: BytesLimitType => blt.size.evaluateIntConst match {
-        case Some(x) => FixedSized(x)
+        case Some(x) => FixedSized(x.toInt)
         case None => DynamicSized
       }
       case _: BytesTerminatedType => DynamicSized
